+++ examples/advanced/10-constructor/reconciler/pipeline_reconciler.go
@@ -42,7 +42,7 @@ import (
 )
 
 const (
-	finalizerName = "orkestra.konductor.io/pipeline-cleanup"
+	finalizerName = "orkestra.orkspace.io/pipeline-cleanup"
 	backoffLimit  = 5
 )